requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2              # via pip-tools
mouseinfo==0.1.3          # via pyautogui
pillow==8.0.1             # via -r requirements.in, mouseinfo, pyscreeze
pip-tools==5.3.1          # via -r requirements.in
pyautogui==0.9.52         # via -r requirements.in
pygetwindow==0.0.9        # via pyautogui
pymsgbox==1.0.9           # via pyautogui
pyperclip==1.8.1          # via mouseinfo
pyrect==0.1.4             # via pygetwindow
pyscreeze==0.1.26         # via pyautogui
python3-xlib==0.15        # via mouseinfo, pyautogui
pytweening==1.0.3         # via pyautogui
six==1.15.0               # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
